+++ lib/stagehand.dart
@@ -26,11 +26,9 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'generators/console_full.dart';
-import 'generators/console_simple.dart';
 import 'generators/package_simple.dart';
 import 'generators/server_shelf.dart';
 import 'generators/web_angular.dart';
-import 'generators/web_angular_simple.dart';
 import 'generators/web_simple.dart';
 import 'generators/web_stagexl.dart';
 import 'src/common.dart';
@@ -38,11 +36,9 @@ import 'src/common.dart';
 /// A curated, prescriptive list of Dart project generators.
 final List<Generator> generators = [
   new ConsoleFullGenerator(),
-  new ConsoleSimpleGenerator(),
   new PackageSimpleGenerator(),
   new ServerShelfGenerator(),
   new WebAngularGenerator(),
-  new WebAngularSimpleGenerator(),
   new WebSimpleGenerator(),
   new WebStageXlGenerator()
 ]..sort();