+++ templates/console-simple/bin/main.dart
@@ -1,3 +1,0 @@
-main(List<String> args) {
-  print('Hello world!');
-}